Disable Hashie::Mash warnings. (#26)
We get many hundreds of lines of warning of this type in some of our projects:
"WARN -- : You are setting a key that conflicts with a built-in method Orchparty::AST::Node#size defined in Hash. This can cause unexpected behavior when accessing the key as a property. You can still access the key via the #[] method."